Adding new test for no BOM and renaming a few tests
